Added graph display functionality for comment analysis
Added functionality to display a graph of the user's comment activity over the specified time period. The graph displays the number of comments and upvotes as lines in red and blue respectively, with days on the y-axis. The plot is saved as a file and added to the reply to the user.